add Auth

+++ app/Http/Controllers/PostController.php
@@ -10,6 +10,11 @@ use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -55,7 +60,7 @@ class PostController extends Controller
         $post->title = $request->title;
         $post->short_title = Str::length($request->title)>30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
         $post->descr = $request->descr;
-        $post->author_id = rand(1, 4);
+        $post->author_id = \Auth::user()->id;
 
         if ($request->file('img')){
             $path = Storage::putFile('public', $request->file('img'));
@@ -77,6 +82,9 @@ class PostController extends Controller
     {
         $post =Post::join('users', 'author_id', '=', 'users.id')
             ->find($id);
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Errors');
+        }
 
         return view('posts.show', compact('post'));
     }
@@ -90,6 +98,15 @@ class PostController extends Controller
     public function edit($id)
     {
         $post =Post::find($id);
+
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Errors');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('You can not edit');
+        }
+
         return view('posts.edit', compact('post'));
 
     }
@@ -104,6 +121,14 @@ class PostController extends Controller
     public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Errors');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('You can not edit');
+        }
         $post->title = $request->title;
         $post->short_title = Str::length($request->title)>30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
         $post->descr = $request->descr;
@@ -128,6 +153,15 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Errors');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('You can not delete');
+        }
+
         $post->delete();
         return redirect()->route('post.index')->with('success', 'Пост успешно удален!');
     }
